have new attributes working, need to get defaults, no uuid, and asociations next
