+++ lib/adapters/ets/helpers.ex
@@ -21,10 +21,10 @@ defmodule ActiveMemory.Adapters.Ets.Helpers do
   end
 
   def to_struct(tuple, module) when is_tuple(tuple),
-    do: struct(module, build_struct(module.__meta__.attributes, tuple))
+    do: struct(module, build_struct(module.__attributes__(:query_fields), tuple))
 
   def to_tuple(%{__struct__: module} = struct) do
-    module.__meta__.attributes
+    module.__attributes__(:query_fields)
     |> Enum.into([], fn key -> Map.get(struct, key) end)
     |> List.to_tuple()
   end
