+++ lib/adapters/mnesia/mnesia.ex
@@ -17,8 +17,8 @@ defmodule ActiveMemory.Adapters.Mnesia do
 
   def create_table(table, _options) do
     options =
-      [attributes: table.__meta__.attributes]
-      |> Keyword.merge(table.__meta__.table_options)
+      [attributes: table.__attributes__(:query_fields)]
+      |> Keyword.merge(table.__attributes__(:table_options))
 
     case :mnesia.create_table(table, options) do
       {:atomic, :ok} ->
@@ -135,8 +135,8 @@ defmodule ActiveMemory.Adapters.Mnesia do
   end
 
   defp build_mnesia_match_spec(query, table) do
-    query_map = :erlang.apply(table, :__meta__, []) |> Map.get(:query_map)
-    match_head = :erlang.apply(table, :__meta__, []) |> Map.get(:match_head)
+    query_map = :erlang.apply(table, :__attributes__, [:query_map])
+    match_head = :erlang.apply(table, :__attributes__, [:match_head])
 
     MatchSpec.build(query, query_map, match_head)
   end
