+++ mix.exs
@@ -45,6 +45,7 @@ defmodule ActiveMemory.MixProject do
   # Run "mix help deps" to learn about dependencies.
   defp deps do
     [
+      {:uuid, "~> 1.1"},
       {:ex_doc, "~> 0.27", only: :dev, runtime: false}
     ]
   end
