+++ test/support/dogs/dog.ex
@@ -1,6 +1,14 @@
 defmodule Test.Support.Dogs.Dog do
   use ActiveMemory.Table,
-    attributes: [:name, :breed, :weight, :dob, fixed?: true, nested: %{one: nil, default: true}],
     type: :ets,
     options: [compressed: true, read_concurrency: true]
+
+  attributes do
+    field(:name, :string)
+    field(:breed, :string)
+    field(:weight, :integer)
+    field(:dob, :string)
+    field(:fixed?, :boolean, default: true)
+    field(:nested, :map)
+  end
 end
