+++ test/support/people/person.ex
@@ -1,5 +1,13 @@
 defmodule Test.Support.People.Person do
   use ActiveMemory.Table,
-    attributes: [:email, :first, :last, :hair_color, :age, :cylon?],
     options: [index: [:last, :cylon?]]
+
+  attributes do
+    field(:email, :string)
+    field(:first, :string)
+    field(:last, :string)
+    field(:hair_color, :string)
+    field(:age, :integer)
+    field(:cylon?, :boolean)
+  end
 end
